+++ noodles-cram/src/async.rs
@@ -1,3 +1,4 @@
 //! Async CRAM.
 
+pub mod fs;
 pub mod io;